+++ src/main/java/com/kpei/mkbd/datatransform/dto/Vd57bDto.java
@@ -2,14 +2,16 @@ package com.kpei.mkbd.datatransform.dto;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Setter
 @Getter
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
 public class Vd57bDto extends Vd5Base {
-    private double saldo;
-    private double dimiliki;
-    private double dipisahkan;
-    private double tidakDipisahkan;
+    private BigDecimal saldo;
+    private BigDecimal dimiliki;
+    private BigDecimal dipisahkan;
+    private BigDecimal tidakDipisahkan;
 }
